requirements.txt
flask
pymysql
flask-sqlalchemy
flask-script
flask-wtf
flask-migrate
py-bcrypt
python-slugify
flask-markdown
flask-uploads
beautifulsoup4
lxml
html5lib
requests
feedparser
